+++ src/main/java/com/dasom/whick/entity/Pressure.java
@@ -14,6 +14,6 @@ public class Pressure {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Integer pressure;
+    private Float pressure;
     private String status;
 }
